+++ v6/src/utils/memory.hpp
@@ -1,6 +1,6 @@
 #include <sys/sysinfo.h>
 
-std::pair<long long int, long int> get_mem_usage_and_free_mem() {
+std::pair<long long int, long long int> get_mem_usage_and_free_mem() {
 	struct sysinfo info;
 	if (sysinfo(&info) < 0)
 		throw;
